pyproject.toml
[tool.poetry]
name = "jinair"
version = "0.1.0"
description = ""
authors = ["wjlee <wjlee.ling@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = "^0.2.6"
langchain-openai = "^0.1.14"
streamlit = "^1.36.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.6"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
